Do some basic refactoring, splitting JavaPlatform into AWT and Android.
Also deleted the OutputFolderName property in AbstractPlatform because it isn't used anymore.
